Rename Sample App
*Rename sample app from Sample to Recent Images to be more descriptive.
