+++ recentimages/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "25.0.2"
 
     defaultConfig {
-        minSdkVersion 9
+        minSdkVersion 10
         targetSdkVersion 25
         versionCode 7
         versionName "1.1.1"
