+++ sample/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.amirarcane.sample"
+<manifest package="com.amirarcane.recentimagesapp"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
 	<uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
