+++ sample/src/main/java/com/amirarcane/recentimagesapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.amirarcane.sample;
+package com.amirarcane.recentimagesapp;
 
 import android.Manifest;
 import android.app.Activity;
