+++ sample/src/main/java/com/amirarcane/recentimagesapp/RecyclerItemClickListener.java
@@ -1,4 +1,4 @@
-package com.amirarcane.sample;
+package com.amirarcane.recentimagesapp;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
