+++ sample/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-	<string name="app_name">Sample</string>
+	<string name="app_name">Recent Images</string>
 
 	<string name="hello_world">Hello world!</string>
 	<string name="action_settings">Settings</string>
